gcpfunctions/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.*
python-dotenv==1.*